Update BPARules-PowerBI.json
Added 'or Description.IndexOf("TODO", StringComparison.OrdinalIgnoreCase) >= 0' to the TODO rule. 

Currently in T/E, changes to the DAX in a calculated column are not permissible, but changes to comments are, so when users are going through a model that has a bunch of unnecessary calculated columns and replacing with measures. you can flag them in T/E with a TODO flag in the Description section then clean them up as the project progresses.